Rename USB serial files to match with library name
