+++ libraries/USBSerial/USBSerial.h
